// src/common/constants/validation-codes.enum.ts
export enum ValidationCode {
  USERNAME_REQUIRED = 'USERNAME_REQUIRED',
  USERNAME_MIN = 'USERNAME_MIN',
  USERNAME_MAX = 'USERNAME_MAX',
  USERNAME_PATTERN = 'USERNAME_PATTERN',
  USERNAME_UNIQUE = 'USERNAME_UNIQUE',

  BIRTHDATE_REQUIRED = 'BIRTHDATE_REQUIRED',
  BIRTHDATE_ISO = 'BIRTHDATE_ISO',
  BIRTHDATE_INVALID = 'BIRTHDATE_INVALID',

  EMAIL_REQUIRED = 'EMAIL_REQUIRED',
  EMAIL_INVALID = 'EMAIL_INVALID',

  PASSWORD_REQUIRED = 'PASSWORD_REQUIRED',
  PASSWORD_WEAK = 'PASSWORD_WEAK',

  VALIDATION_UNKNOWN = 'VALIDATION_UNKNOWN',
}